agent/requirements.txt
pydantic-ai-slim[ag-ui] 
uvicorn
dotenv